refactored archive into gem project
