+++ .gitignore
@@ -1,3 +1,4 @@
 /downloads
+/lib/downloads
 /spec/tmp
 .env
