--- archive/Gemfile
@@ -1,4 +1,0 @@
-source "http://rubygems.org"
-
-gem "vine_client"
-gem "dotenv"
