--- archive/lib/vine_dl.rb
@@ -1,5 +1,0 @@
-module VineDl
-end
-
-require_relative "vine_dl/user"
-require_relative "vine_dl/video"
