+++ lib/vine_dl/user.rb
