+++ lib/vine_dl/video.rb
