+++ spec/mocks/client_mock.rb
