+++ spec/mocks/client_video_mock.rb
