+++ spec/mocks/user_mock.rb
