+++ spec/mocks/video_mock.rb
