+++ spec/support/client_video_factory.rb
