--- spec/vine_dl_spec.rb
@@ -1,11 +1,0 @@
-require 'spec_helper'
-
-describe VineDl do
-  it 'has a version number' do
-    expect(VineDl::VERSION).not_to be nil
-  end
-
-  it 'does something useful' do
-    expect(false).to eq(true)
-  end
-end
