+++ vine_dl.gemspec
@@ -19,6 +19,13 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
 
+  spec.add_development_dependency "hashie", ">= 2.0.2", "< 3.0.0"
   spec.add_development_dependency "bundler", "~> 1.12"
   spec.add_development_dependency "rake", "~> 10.0"
+  spec.add_development_dependency "vine_client", "~> 0.0.7"
+  spec.add_development_dependency "minitest"
+  spec.add_development_dependency "dotenv", "~> 2.1.1"
+
+
+  spec.add_runtime_dependency "vine_client", "~> 0.0.7"
 end
